Bug 1971425 - Don't start a Nimbus update during shutdown r=nimbus-reviewers,relud
We used to prevent new updates from happening via the Normandy
CleanupManager, which adds shutdown blockers at the
`profile-before-change` phase. This phase is quite late in terms of
shutdown. If we know we are shutting down, we should try to minimize the
amount of work we're doing and not let new updates trigger.

We now disable the RemoteSettingsExperimentLoader at the earliest part
of shutdown: `app-shutdown-confirmed`. Any pending update will finish
before we unblock shutdown.

Differential Revision: https://phabricator.services.mozilla.com/D253731

commit-id:355cdc2b